requirements.txt
duckduckgo-search==6.3.5
playwright==1.48.0
playwright-stealth==1.0.6
google-generativeai==0.8.3
pandas==2.2.3
tenacity==9.0.0
colorama==0.4.6
tldextract==5.1.3
tqdm==4.67.1
Pillow==11.0.0

# Gmail API packages for email automation
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

